requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
Pillow>=9.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
WTForms>=3.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-flask>=1.2.0